Apply custom CSS during build
Prior to this commit the `index.html` was simply
copied into the build to apply custom CSS styles.
However this caused an issue where newer versions
of HaxeFlixel would invoke the game using
a differently Javascript method.

To prevent these updates from breaking the game,
the CSS styles are simply added into the `head`
of the HTML file after the build.